Updated the PR based on the comments from Carlos and Patricia
